+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
+## 4.18.0
+
+The Sentry SDK team is happy to announce the immediate availability of Sentry PHP SDK v4.18.0.
+
+### Features
+
+- Add support for feature flags. [(#1951)](https://github.com/getsentry/sentry-php/pull/1951)
+```php
+\Sentry\SentrySdk::getCurrentHub()->configureScope(function (\Sentry\State\Scope $scope) {
+    $scope->addFeatureFlag("my.feature.enabled", true);
+});
+```
+- Add more representations for log attributes instead of dropping them. [(#1950)](https://github.com/getsentry/sentry-php/pull/1950)
+
+### Misc
+
+- Merge log attributes in a separate method. [(#1931)](https://github.com/getsentry/sentry-php/pull/1931)
+
 ## 4.17.1
 
 The Sentry SDK team is happy to announce the immediate availability of Sentry PHP SDK v4.17.1.